autonix-ai-edge/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
flask>=3.0.0
pyserial>=3.5
ultralytics>=8.0.0
colorama>=0.4.6
